refactor: consolidate lazy getattr into getattr module
